Extract `FeedCaheSpy` into a shared scope to remove duplication.
