+++ EssentialFeed/EssentialFeedTests/Feed Cache/Helpers/FeedStoreSpy.swift
@@ -1,0 +1,52 @@
+//
+//  FeedStoreSpy.swift
+//  EssentialFeed
+//
+//  Created by Octavio Rojas on 30/10/24.
+//
+
+import Foundation
+import EssentialFeed
+
+class FeedStoreSpy: FeedStore {
+    enum ReceivedMessage: Equatable {
+        case deleteCachedFeed
+        case insert([LocalFeedImage], Date)
+        case retrieve
+    }
+    
+    private(set) var receivedMessages = [ReceivedMessage]()
+    
+    private var deletionCompletions = [DeletionCompletion]()
+    private var insertionCompletions = [InsertionCompletion]()
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        deletionCompletions.append(completion)
+        receivedMessages.append(.deleteCachedFeed)
+    }
+    
+    func completeDeletion(with error: Error?, at index: Int = 0) {
+        deletionCompletions[index](error)
+    }
+    
+    func completeDeletionSuccessfully(at index: Int = 0) {
+        deletionCompletions[index](nil)
+    }
+    
+    func insert(_ feed: [LocalFeedImage], timeStamp: Date, completion: @escaping InsertionCompletion) {
+        insertionCompletions.append(completion)
+        receivedMessages.append(.insert(feed, timeStamp))
+    }
+    
+    func completeInsertion(with error: Error?, at index: Int = 0) {
+        insertionCompletions[index](error)
+    }
+    
+    func completeInsertionSuccessfully(at index: Int = 0) {
+        insertionCompletions[index](nil)
+    }
+    
+    func retrieve() {
+        receivedMessages.append(.retrieve)
+    }
+}
